backend_final_project/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
psycopg==3.1.12
pytz==2023.3.post1
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.3
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0